[BUGFIX] Remove paste icons in TYPO3 8.7
As there is no way to append own parameter to DataHandler actions, it
is not possible to assign the correct container pid for paste icon links.
Therefore the paste icons are removed with this patch. You can still use
the context menu to paste a copied record.

Furthermore this patch adds an addition "Content" button underneath a
record element.